add secondary kernel headers and glibc installation
The basic install for the kernel headers and glibc are done to $SYSROOT,
using the normal assumption that the sysroot does in fact contain a subset
of the files that belong in the root filesystem of the target device.  This
build additionally installs the same kernel headers, and the same version
of glibc, with a prefix of /tools.  If we link everything against that
glibc, I'm hopeful that it will build everything to be self-contained
within the /tools directory of the final system (just like in LFS)