restructure the automation files
I'm still not especially certain that I like this structure for
the "literate programming" style of build automation.  Still, I
think this is a reasonable way of proceeding:  write the
canonical material that will be used for both book and automated
build scripts, and that can form the acceptance test for the
conversion programs (the equivalent of weave and tangle).