Lots of flailing about.
Although we can get the drive to spin, we can't seem to get the
hard disk to be recognized.

Using 2.6.23, when we load the aec62xx module we get a bus error
and kernel panic.

We moved up to 2.6.28.6 (OpenWRT svn revision 14587), but then
insmodding aec62xx doesn't do anything at all.  Strangely, the
PCI device for the IDE controller seems to be 1191:0010, rather
than 1191:0008 (which the kernel expects).