The initial glibc build can't use stack-protector
The initial GCC doesn't support -fstack-protector, but it's
detected as working by the autoconf script for some reason. Maybe
because the configure scripts are run against the host GCC rather
than the partial GCC, or something like that.

So tell the configure machinery that it shouldn't check and SSP
is not available.